### 2025-06-06 #8359 @editorialbot
[PRE REVIEW]: TMLE.jl: Targeted Minimum Loss-Based Estimation In Julia.
**Submitting author:** @olivierlabayle (<a href="http://orcid.org/0000-0002-3708-3706">Olivier Labayle</a>)
**Repository:** https://github.com/TARGENE/TMLE.jl
**Branch with paper.md** (empty if default branch): main
**Version:** v0.19.0
