### 2025-06-06 268ff28
docstrings and generalization in lasso strategy
