### 2025-06-06 85ac63c
add some more dev docs
