+++ previews/PR135/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"examples/interactions_correlated/#Interaction-Estimation","page":"Interaction Estimation","title":"Interaction Estimation","text":"","category":"section"},{"location":"examples/interactions_correlated/","page":"Interaction Estimation","title":"Interaction Estimation","text":"In this example we aim to estimate the average interaction effect of two, potentially correlated, treatment variables T1 and T2 on an outcome Y.","category":"page"},{"location":"examples/interactions_correlated/#Data-Generating-Process","page":"Interaction Estimation","title":"Data Generating Process","text":"","category":"section"},{"location":"examples/interactions_correlated/","page":"Interaction Estimation","title":"Interaction Estimation","text":"Let's consider the following structural causal model where the shaded nodes represent the observed variables.","category":"page"},{"location":"examples/interactions_correlated/","page":"Interaction Estimation","title":"Interaction Estimation","text":"(Image: interaction-graph)","category":"page"},{"location":"examples/interactions_correlated/","page":"Interaction Estimation","title":"Interaction Estimation","text":"In other words, only one confounding variable is observed (W1). This would be a major problem if we wanted to estimate the average treatment effect of T1 or T2 on Y separately. However, here, we are interested in interactions and thus W1 is a sufficient adjustment set. This artificial situation is ubiquitous in genetics, where two main sources of confounding exist. Ancestry, can be estimated (here W1) and linkage disequilibrium is usually more challenging to address (here W2).","category":"page"},{"location":"examples/interactions_correlated/","page":"Interaction Estimation","title":"Interaction Estimation","text":"Let us first define some helper functions and import some libraries.","category":"page"},{"location":"examples/interactions_correlated/","page":"Interaction Estimation","title":"Interaction Estimation","text":"using Distributions\nusing Random\nusing DataFrames\nusing Statistics\nusing CategoricalArrays\nusing TMLE\nusing CairoMakie\nusing MLJXGBoostInterface\nusing MLJBase\nusing MLJLinearModels\nusing MLJTuning\nusing StatisticalMeasures\n\nfunction estimate_across_correlation_levels(estimators, σs; n=1000)\n    results = Dict(key => [] for key in keys(estimators))\n    for σ in σs\n        dataset = generate_dataset(n=n, σ=σ)\n        for (estimator_key, estimator) in estimators\n            result, _ = estimator(Ψ, dataset; verbosity=0)\n            push!(results[estimator_key], result)\n        end\n    end\n    return results\nend\n\nfunction estimate_across_sample_sizes_and_correlation_levels(estimators, ns, σs)\n    results = []\n    for n in ns\n        results_at_n = estimate_across_correlation_levels(estimators, σs; n=n)\n        push!(results, results_at_n)\n    end\n    return results\nend\n\nfunction plot_across_sample_sizes_and_correlation_levels(results, ns, σs; estimator=\"TMLE_SL\", title=\"Estimation via TMLE (GLMs)\")\n    fig = Figure(size=(800, 800))\n    for (index, n) in enumerate(ns)\n        results_at_n = results[index][estimator]\n        Ψ̂s = TMLE.estimate.(results_at_n)\n        errors = last.(confint.(significance_test.(results_at_n))) .- Ψ̂s\n        ax = if n == last(ns)\n            Axis(fig[index, 1], xlabel=\"σ\", ylabel=\"AIE\\n(n=$n)\")\n        else\n            Axis(fig[index, 1], ylabel=\"AIE\\n(n=$n)\", xticklabelsvisible=false)\n        end\n        errorbars
[…]
