+++ docs/src/contributing.md
@@ -16,7 +16,7 @@ Here is some general information about the code base.
 
 Implementing a new estimand requires both the definition of the estimand (see `src/counterfactual_mean_based/estimands.jl` for examples) and the associated estimators.
 
-The entry point estimators are defined within `src/counterfactual_mean_based/estimators.jl`. This is because there has been a focus on implementationsestimands based on the counterfactual mean ``E[Y(T)]``. However the estimation procedures are still quite general and could serve as a backbone for future estimands. For instance:
+Currently, the entry point estimators are defined within `src/counterfactual_mean_based/estimators.jl`. This is because there has been a focus on implementating estimands based on the counterfactual mean ``E[Y(T)]``. However the estimation procedures are still quite general and could serve as a backbone for future estimands. For instance:
 
 - `get_relevant_factors`: returns the nuisance parameters for a given parameter ``Ψ``.
 - `gradient_and_estimate`: computes the gradient for a given parameter ``Ψ``.
@@ -28,7 +28,7 @@ There are several interesting directions to complement this package with new est
 - C-TMLE: collaborative TMLE is a powerful estimation strategy for which a rough template is in place (`src/counterfactual_mean_based/collaborative_template.jl`). Many strategies following the template can then be implemented. Get in touch if needed for general directions.
 - Riesz Representer Estimation: Estimation of the nuisance functions is typically made by estimation of the propensity score. In cases of positivity violations this can lead to numerical instability. It turns out that the inverse of the propensity score can be estimated directly (see [here](https://github.com/TARGENE/TMLE.jl/issues/83)).
 
-### General Code Pattern
+### Estimand - Estimator - Estimate
 
 The package is centered around the following statistical concepts which are embodied by [Julia structs](https://docs.julialang.org/en/v1/manual/types/):
 
