### 2025-06-19 29b7782
add lasso_strategy
