+++ src/counterfactual_mean_based/lasso_strategy.jl
@@ -1,0 +1,211 @@
+##################################################################
+###                      LassoStrategy                         ###
+##################################################################
+
+using GLMNet
+using DataFrames
+using StatsFuns
+using ..TMLE
+using Statistics
+
+"""
+    LassoCTMLE <: CollaborativeStrategy
+
+Lasso-based Collaborative TMLE strategy using cross-validated lambda selection.
+"""
+struct LassoCTMLE <: CollaborativeStrategy
+    lambda_path::Vector{Float64}
+    cv_folds::Int
+    best_lambda::Union{Nothing,Float64}
+    losses::Vector{Float64}
+    confounders::Vector{Symbol}
+    patience::Int
+    function LassoCTMLE(;
+        lambda_path = exp10.(range(-4, stop = 0, length = 50)),
+        cv_folds = 5,
+        confounders = Symbol[],
+        patience = 3,
+    )
+        isempty(confounders) &&
+            throw(ArgumentError("Must specify confounders for LassoCTMLE"))
+        new(lambda_path, cv_folds, nothing, Float64[], confounders, patience)
+    end
+end
+
+"""
+    stratified_kfold(y::AbstractVector, k::Int)
+
+Returns a vector of (train_idx, test_idx) tuples for stratified k-fold cross-validation,
+where `y` contains class labels.
+"""
+function stratified_kfold(y::AbstractVector, k::Int)
+    idx_by_class = Dict{eltype(y), Vector{Int}}()
+    for (i, label) in enumerate(y)
+        push!(get!(idx_by_class, label, Int[]), i)
+    end
+
+    folds = [Int[] for _ in 1:k]
+    for idxs in values(idx_by_class)
+        shuffled = shuffle(idxs)
+        for (i, idx) in enumerate(shuffled)
+            push!(folds[mod1(i, k)], idx)
+        end
+    end
+
+    result = Vector{Tuple{Vector{Int}, Vector{Int}}}(undef, k)
+    all_idxs = collect(1:length(y))
+    for i in 1:k
+        test_idx = folds[i]
+        train_idx = setdiff(all_idxs, test_idx)
+        result[i] = (train_idx, test_idx)
+    end
+    return result
+end
+
+function initialise!(strategy::LassoCTMLE, Ψ)
+    strategy.best_lambda = nothing
+    empty!(strategy.losses)
+    return nothing
+end
+
+function update!(strategy::LassoCTMLE, last_targeted_η̂ₙ, dataset)
+    return nothing
+end
+
+finalise!(strategy::LassoCTMLE) = nothing
+
+function exhausted(strategy::LassoCTMLE)
+    strategy.best_lambda !== nothing
+end
+
+"""
+    propensity_score(Ψ, strategy::LassoCTMLE)
+
+Returns a function which, given a dataset, fits LASSO at the chosen lambda and predicts propensity scores.
+This method is generic: it works for any valid Ψ parameter by using TMLE.treatment(Ψ).
+"""
+function propensity_score(Ψ, strategy::LassoCTMLE)
+    function fit_predict(dataset; lambda_override = nothing)
+        W = Matrix(select(dataset, strategy.confounders))
+        A = dataset[!, treatment(Ψ)]
+        λ = isnothing(lambda_override) ? [strategy.best_lambda] : [lambda_override]
+        g_fit = glmnet(W, A, Binomial(), lambda = λ)
+        g_pred = GLMNet.predict(g_fit, W, lambda = λ[1])
+        clamp.(g_pred, 0.01, 0.99) # stabilize
+    end
+    return fit_predict
+end
+
+"""
[…]
