### 2025-06-22 #8359 @jbytecode
@crvernon - sure!
