### 2025-06-22 #8359 @jbytecode
@olivierlabayle - Hi, thanks for submitting your work to JOSS; I'm the handling editor for this submission.

I'm now inviting nhejazi as a reviewer. I had also considered VasanthRajendran, but they haven't responded to previous invitations for other submissions. If you have additional reviewer suggestions, please let me know.

Thank you in advance.
