### 2025-06-22 #8359 @jbytecode
👋👋👋 Dear @mantzaris 👋👋👋

**This submission is about statistics and Julia, so it fits your expertise well.**

**[[PRE REVIEW]: TMLE.jl: Targeted Minimum Loss-Based Estimation In Julia.](https://github.com/openjournals/joss-reviews/issues/8359)**

You can find more information at the top of this Github issue ([LINK](https://github.com/openjournals/joss-reviews/issues/8359)).

Thank you in advance!
