### 2025-06-22 #8359 @nhejazi
hi @jbytecode — thanks for tagging me in for this review. i’d be happy to help, but would it be alright if i were to co-review this with @salbalkus? i’m fluent in TMLE but less so in Julia, and we would be able to work in the review together. @salbalkus is the primary author of the JOSS-reviewed https://github.com/salbalkus/CausalTables.jl
