### 2025-06-22 #8359 @salbalkus
@jbytecode @nhejazi Yes, I would be willing to assist. I will disclose, though: I've contributed to the TMLE.jl repo, but my contribution has been limited to integrating the package with my own package CausalTables.jl (not to the paper itself). If this isn't too much of a conflict of interest, then I can certainly review.
