### 2025-06-22 #8359 @jbytecode
@salbalkus - Thank you for your reply and  the clarification. I've reviewed your contributions at https://github.com/TARGENE/TMLE.jl/commits?author=salbalkus. While these commits are very useful, they don't appear to introduce significant new functionality. Therefore, I don't believe this constitutes a conflict of interest.
