### 2025-06-22 #8359 @mantzaris
@jbytecode I am happy to review, do you still want/need my review?
