### 2025-06-22 #8359 @jbytecode
@mantzaris - yes, sure, thank you for accepting.

I'm now starting the review. A new thread will be created. I'll introduce the instructions, see you there!
