### 2025-06-22 #8446 @editorialbot
[REVIEW]: TMLE.jl: Targeted Minimum Loss-Based Estimation in Julia
**Submitting author:** @olivierlabayle (<a href="http://orcid.org/0000-0002-3708-3706">Olivier Labayle</a>)
**Repository:** https://github.com/TARGENE/TMLE.jl
**Branch with paper.md** (empty if default branch): ol_dev_doc
**Version:** v0.20.2
