### 2025-06-23 #8446 @olivierlabayle
@mantzaris thank you for reporting this. Could you let me know which version of the package is installed? I assume `using TMLE` worked fine. Locally I get:

```julia
(temp) pkg> st
Status `~/Dev/TMLE.jl/temp/Project.toml`
  [8afdd2fb] TMLE v0.19.0

julia> using TMLE
[ Info: Precompiling ParsersExt [f526588d-e68b-5dc5-a62e-ff9f36e48b1a] (cache misses: wrong dep version loaded (6), incompatible header (10), mismatched flags (4))
[ Info: Precompiling JSONExt [6b5134e7-a302-5195-a64b-aad213349502] (cache misses: wrong dep version loaded (8), mismatched flags (10))

julia> Tmle()
Tmle(Dict{Symbol, MLJBase.ProbabilisticPipeline{N, MLJModelInterface.predict} where N<:NamedTuple}(:Q_binary_default => ProbabilisticPipeline(continuous_encoder = ContinuousEncoder(drop_last = true, …), …), :G_default => ProbabilisticPipeline(continuous_encoder = ContinuousEncoder(drop_last = true, …), …), :Q_continuous_default => ProbabilisticPipeline(continuous_encoder = ContinuousEncoder(drop_last = true, …), …)), nothing, nothing, 1.0e-8, true, nothing, 1, false)
```
