### 2025-06-23 #8446 @olivierlabayle
By the normal environment I assume you mean the based environment. I suspect some other installed package in this environment is restricting compatibility and forcing 0.18 for which you can find the docs here: https://targene.github.io/TMLE.jl/v0.18/. The docs are all versioned.
