### 2025-06-23 #8446 @jbytecode
I clone the repository in my temp directory of Linux installation using

`> git clone https://github.com/TARGENE/TMLE.jl.git
`

and then `cd` to the folder and started Julia using

`> julia --project=. -t auto`

After performing

`] up`

and

`] test`

I got

```julia
331.818206 seconds (1.23 G allocations: 79.352 GiB, 3.34% gc time, 50 lock conflicts, 84.84% compilation time: 8% of which was recompilation)
     Testing TMLE tests passed
```

Since the package depends on a tons of packages, installation in the root directory may cause version conflicts.

@mantzaris - Which packages are installed in your root environment of the Julia installation? Could you please provide the error messages if possible?

@olivierlabayle - You may want to reduce the dependencies if possible.
