### 2025-06-24 7b8252e
remove some deps
