+++ src/counterfactual_mean_based/nuisance_estimators.jl
@@ -80,19 +80,20 @@ end
 @auto_hash_equals struct CMRelevantFactorsEstimator <: Estimator
     train_validation_indices
     models::Dict
+    prevalence::Union{Nothing, Float64}
 end
 
-CMRelevantFactorsEstimator(;models, train_validation_indices=nothing) = CMRelevantFactorsEstimator(train_validation_indices, models)
+CMRelevantFactorsEstimator(;models, train_validation_indices=nothing, prevalence=nothing) = CMRelevantFactorsEstimator(train_validation_indices, models, prevalence)
 
 """
 If there is no collaborative strategy, we are in CV mode and `train_validation_indices` are used to build the initial estimator.
 """
-CMRelevantFactorsEstimator(collaborative_strategy::Nothing; models, train_validation_indices=nothing) = CMRelevantFactorsEstimator(train_validation_indices, models)
+CMRelevantFactorsEstimator(collaborative_strategy::Nothing; models, train_validation_indices=nothing, prevalence=nothing) = CMRelevantFactorsEstimator(train_validation_indices, models, prevalence)
 
 """
 If there is a collaborative strategy, `train_validation_indices` are ignored to build the initial estimator.
 """
-CMRelevantFactorsEstimator(collaborative_strategy; models, train_validation_indices=nothing) = CMRelevantFactorsEstimator(nothing, models)
+CMRelevantFactorsEstimator(collaborative_strategy; models, train_validation_indices=nothing, prevalance=nothing) = CMRelevantFactorsEstimator(nothing, models, prevalance)
 
 function acquire_model(models, key, dataset, is_propensity_score)
     # If the model is in models return it
@@ -108,12 +109,13 @@ end
 
 function build_propensity_score_estimator(propensity_score, models, dataset;
     train_validation_indices=nothing,
+    prevalence=nothing
     )
     cd_estimators = Dict()
     for conditional_distribution in propensity_score
         outcome = conditional_distribution.outcome
         model = acquire_model(models, outcome, dataset, true)
-        cd_estimators[outcome] = ConditionalDistributionEstimator(model, train_validation_indices)
+        cd_estimators[outcome] = ConditionalDistributionEstimator(model, train_validation_indices, prevalence=prevalence)
     end
     return JointConditionalDistributionEstimator(cd_estimators)
 end
@@ -123,13 +125,15 @@ function estimate_propensity_score(propensity_score, models, dataset;
     cache=Dict(),
     verbosity=1,
     machine_cache=false,
-    acceleration=CPU1()
+    acceleration=CPU1(),
+    prevalence=nothing
     )
     propensity_score_estimator = build_propensity_score_estimator(
         propensity_score, 
         models,  
         dataset;
         train_validation_indices=train_validation_indices,
+        prevalence=prevalence
     )
     return propensity_score_estimator(
         propensity_score, 
@@ -146,12 +150,14 @@ function estimate_outcome_mean(outcome_mean, models, dataset;
     cache=Dict(),
     verbosity=1,
     machine_cache=false,
-    acceleration=CPU1()
+    acceleration=CPU1(),
+    prevalence=nothing
     )
     outcome_model = acquire_model(models, outcome_mean.outcome, dataset, false)
     outcome_mean_estimator = ConditionalDistributionEstimator(
         outcome_model,
-        train_validation_indices, 
+        train_validation_indices,
+        prevalence=prevalence 
     )
     return try_fit_ml_estimator(outcome_mean_estimator, outcome_mean, dataset;
         error_fn=outcome_mean_fit_error_msg,
@@ -171,20 +177,23 @@ function estimate_propensity_score_and_outcome_mean(
[…]

[…]
