### 2025-06-24 f69badc
nuisance functions update; removing problematic package
