+++ Project.toml
@@ -26,7 +26,6 @@ Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 SplitApplyCombine = "03a91e81-4c3e-53e1-a0a4-9c0c8f19dd66"
 StatisticalMeasures = "a19d573c-0a75-4610-95b3-7071388c7541"
-Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 TableOperations = "ab02a1b2-a7df-11e8-156e-fb1833f50b87"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
@@ -65,7 +64,6 @@ OrderedCollections = "1.6.3"
 Printf = "1.11.0,1.10.9"
 SplitApplyCombine = "1.2.2"
 StatisticalMeasures = "0.2"
-Statistics = "1.11.1"
 TableOperations = "1.2"
 Tables = "1.6"
 YAML = "0.4.9"
