### 2025-06-25 #8446 @mantzaris
@jbytecode , I have gone over the documentation and read through the methodology which provides good background. It is not really for 'newcomers' but those versed in statistics (subject matter) will find value in it. The api is dense but not cluttered. Each function offers quite a bit in terms of options which users can find useful. The code looks correct, after picking some functions out of the api to see how they work.

requests,
- more tests for `estimands.jl`
- and for 'causaltables_interface.jl' does it suffice that pretty much only the type checking is really done?
- and in the docs for the section [Classic Structural Causal Models](https://targene.github.io/TMLE.jl/stable/user_guide/scm/#Classic-Structural-Causal-Models) some more discussion and examples
