### 2025-06-25 02c68b5
Revert changes to Project.toml and .gitignore
