### 2025-06-27 12b2176
add some more docs for SCM
