+++ docs/src/user_guide/scm.md
@@ -4,7 +4,18 @@ CurrentModule = TMLE
 
 # Structural Causal Models
 
-Even if you don't have to, it can be useful to define a Structural Causal Model (`SCM`) for your problem. A `SCM` is a directed acyclic graph that describes the causal relationships between the random variables under study.
+In TMLE.jl, specifying a structural causal model (`SCM`) is optional but can be useful to reason about your data. It can also be used for causal identification, that is, converting a causal estimand into a statistical estimand which can be estimated from data. A `SCM` is a directed acyclic graph that describes the causal relationships between the random variables under study. However, it does not make any assumption on the functional form of these dependencies. The traditional `SCM` used in most causal inference studies is the following:
+
+![classic_scm](assets/causal_graph.png)
+
+Where:
+
+- ``Y`` is the outcome variable.
+- ``T`` is a set of treatment variables whose effect on ``Y`` is of interest.
+- ``W`` is a set of confounding variables affecting both ``Y`` and ``T``.
+- ``C`` is a set of variables only influencing ``Y``.
+
+We now explain how these models can be built in TMLE.jl.
 
 ## Incremental Construction
 
@@ -15,16 +26,16 @@ using TMLE # hide
 scm = SCM()
 ```
 
-This model does not say anything about the random variables and is thus not really useful. Let's assume that we are interested in an outcome ``Y`` and that this outcome is determined by 8 other random variables. We can add this assumption to the model
+This model does not say anything about the random variables and is thus not very useful. Let's return to the `SCM` presented above and construct it step by step. First let's add the relationship between ``Y`` and its parents:
 
 ```@example scm
-add_equation!(scm, :Y => [:T₁, :T₂, :W₁₁, :W₁₂, :W₂₁, :W₂₂, :W, :C])
+add_equation!(scm, :Y => [:T, :W, :C])
 ```
 
-Let's now assume that we have a more complete knowledge of the problem and we also know how `T₁` and `T₂` depend on the rest of the variables in the system.
+This model does not contain the relationship between ``T`` and ``W``, let's add it now:
 
 ```@example scm
-add_equations!(scm, :T₁ => [:W₁₁, :W₁₂, :W], :T₂ => [:W₂₁, :W₂₂, :W])
+add_equations!(scm, :T => [:W])
 ```
 
 ## One Step Construction
@@ -33,22 +44,37 @@ Instead of constructing the `SCM` incrementally, one can provide all the specifi
 
 ```@example scm
 scm = SCM([
-    :Y  => [:T₁, :T₂, :W₁₁, :W₁₂, :W₂₁, :W₂₂, :W, :C],
-    :T₁ => [:W₁₁, :W₁₂, :W],
-    :T₂ => [:W₂₁, :W₂₂, :W]
+    :Y  => [:T, :W, :C],
+    :T => [:W]
+])
+```
+
+## Multiple Treatments
+
+In many cases, there isn't a single treatment variable but many. Furthermore, each treatment may be confounded by a different set of variables. We can encode this structure within a `SCM`.
+
+```@example scm
+scm = SCM([
+    :Y  => [:T₁, :T₂, :W₁, :W₂, :C],
+    :T₁ => [:W₁],
+    :T₂ => [:W₂]
 ])
 ```
 
 ## Classic Structural Causal Models
 
-There are many cases where we are interested in estimating the causal effect of a some treatment variables on a some outcome variables. If all treatment variables share the same set of confounders, we can quickly define the associated `SCM` with the `StaticSCM` interface:
[…]
