### 2025-06-27 8522386
update image paths
