### 2025-06-30 #8446 @olivierlabayle
I have been addressing the current review comments [here](https://github.com/TARGENE/TMLE.jl/pull/135), the updated docs are also accessible [at this url](https://targene.github.io/TMLE.jl/previews/PR135/).
