+++ src/TMLE.jl
@@ -47,6 +47,7 @@ export Configuration
 export brute_force_ordering, groups_ordering
 export gradients, epsilons, estimates
 export AdaptiveCorrelationStrategy, GreedyStrategy
+export LassoCTMLE
 export CausalStratifiedCV, CV, StratifiedCV, Holdout
 export CPUThreads, CPU1
 
@@ -69,6 +70,7 @@ include("counterfactual_mean_based/fluctuation.jl")
 include("counterfactual_mean_based/collaborative_template.jl")
 include("counterfactual_mean_based/nuisance_estimators.jl")
 include("counterfactual_mean_based/covariate_based_strategies.jl")
+include("counterfactual_mean_based/lasso_strategy.jl")
 include("counterfactual_mean_based/estimators.jl")
 include("counterfactual_mean_based/clever_covariate.jl")
 include("counterfactual_mean_based/gradient.jl")
