+++ src/TMLE.jl
@@ -26,6 +26,8 @@ using DataFrames
 using ComputationalResources
 using Base.Threads
 using Printf
+using GLMNet
+using StatsFuns
 
 # #############################################################################
 # EXPORTS
