+++ Project.toml
@@ -5,7 +5,6 @@ version = "0.19.0"
 
 [deps]
 AutoHashEquals = "15f4f7f2-30c1-5605-9d31-71845cf9641f"
-CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 CategoricalArrays = "324d7699-5711-5eae-9e2f-1d82baa6b597"
 Combinatorics = "861a8166-3701-5b0c-9a16-15d98fcdc6aa"
 ComputationalResources = "ed09eef8-17a6-5b46-8889-db040fac31e3"
@@ -13,9 +12,9 @@ DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 DifferentiationInterface = "a0c0ee7d-e4b9-4e03-894e-1c5f64a51d63"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 GLM = "38e38edf-8417-5370-95a0-9cbb8c7f171a"
+GLMNet = "8d5ece8b-de18-5317-b113-243142960cc6"
 Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
 HypothesisTests = "09f84164-cd44-5f33-b23f-e6b0d136a0d5"
-JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
 LogExpFunctions = "2ab3a3ac-af41-5b50-aa03-7779005ae688"
 MLJBase = "a7f614a8-145f-11e9-1d2a-a57a1082229d"
 MLJGLMInterface = "caf8df21-4939-456d-ac9c-5fefbfb04c0c"
@@ -28,13 +27,15 @@ Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 SplitApplyCombine = "03a91e81-4c3e-53e1-a0a4-9c0c8f19dd66"
 StatisticalMeasures = "a19d573c-0a75-4610-95b3-7071388c7541"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
+StatsFuns = "4c63d2b9-4356-54db-8cca-17b64c39e42c"
 TableOperations = "ab02a1b2-a7df-11e8-156e-fb1833f50b87"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
-YAML = "ddb6d928-2868-570f-bddf-ab3f9cf99eb6"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [weakdeps]
 CausalTables = "6af48e0c-efc2-4bf7-a92f-a553ccf79fd6"
+JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
+YAML = "ddb6d928-2868-570f-bddf-ab3f9cf99eb6"
 
 [extensions]
 CausalTablesExt = "CausalTables"
@@ -43,7 +44,6 @@ YAMLExt = "YAML"
 
 [compat]
 AutoHashEquals = "2.1.0"
-CSV = "0.10.15"
 CategoricalArrays = "0.10"
 CausalTables = "1.2.1"
 Combinatorics = "1.0.2"
@@ -52,6 +52,7 @@ DataFrames = "1.7.0"
 DifferentiationInterface = "0.6.43"
 Distributions = "0.25"
 GLM = "1.8.2"
+GLMNet = "0.7.4"
 Graphs = "1.8"
 HypothesisTests = "0.10, 0.11"
 JSON = "0.21.4"
@@ -65,6 +66,7 @@ OrderedCollections = "1.6.3"
 Printf = "1.11.0,1.10.9"
 SplitApplyCombine = "1.2.2"
[…]
