### 2025-07-02 e3cbe59
add environment section in contributing guide
