### 2025-07-09 #8446 @salbalkus
Hi @jbytecode I still need to advise @nhejazi on the finer points of Julia for this package to inform what more is needed on the paper and documentation in order to complete my review, but so far the package looks good. The tests and most of the code in the docs have run fine for me. A few minor points:

- There appear to be a couple typos preventing the code in both the overview and walkthrough pages from replicating. Specifically, the line `tmle = Tmle()` throws an error where it appears; seems like it should be `tmle = TMLEE()`. Similarly, `ose = Ose()` also yields a similar error; I think it should be `ose = OSE()`
    - Might be worth implementing [doctests](https://documenter.juliadocs.org/stable/man/doctests/) to help catch these types of errors in the future.
- @olivierlabayle is there a reason that the package requires data in the form of a `DataFrame` rather than just a Table? From what I understand, many Julia packages are built around the more general Tables.jl interface, including MLJ on top of which you're building, so I'm curious about that design decision. I think DataFrames.jl might be contributing to some of the previous comments I saw on this thread about large dependencies.
