### 2025-07-09 #8446 @olivierlabayle
Hi @salbalkus

- No `Tmle()` and `Ose()` are correct. This is also likely related to a wrong version being installed as above (please install 0.19 or refer to the versioned documentation for versions <0.19). The docs are already built using the source code so if the CI passes (and it does), then the syntax is correct.
- Yes, this is mostly for ease of data manipulation and in particular because I need to add columns to tables. I have plans to try to make the package work with more general tables but this is out of scope for now. As a comparison, it is important to keep in mind that the standard R package tmle3 depends on data.table, ggplot2 and sl3 (which itself depends on caret).
