+++ Project.toml
@@ -27,9 +27,9 @@ Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 SplitApplyCombine = "03a91e81-4c3e-53e1-a0a4-9c0c8f19dd66"
 StatisticalMeasures = "a19d573c-0a75-4610-95b3-7071388c7541"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
-StatsFuns = "4c63d2b9-4356-54db-8cca-17b64c39e42c"
 TableOperations = "ab02a1b2-a7df-11e8-156e-fb1833f50b87"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
+ToeplitzMatrices = "c751599d-da0a-543b-9d20-d0a503d91d24"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [weakdeps]
@@ -66,9 +66,9 @@ OrderedCollections = "1.6.3"
 Printf = "1.11.0,1.10.9"
 SplitApplyCombine = "1.2.2"
 StatisticalMeasures = "0.2"
-StatsFuns = "1.5.0"
 TableOperations = "1.2"
 Tables = "1.6"
+ToeplitzMatrices = "0.8.5"
 YAML = "0.4.9"
 Zygote = "0.7"
 julia = "1.10, 1"
