### 2025-07-15 b2e28af
case-control weighted fluctuation and gradient; not the desired behavior
