+++ src/counterfactual_mean_based/fluctuation.jl
@@ -6,10 +6,11 @@ mutable struct Fluctuation <: MLJBase.Supervised
     ps_lowerbound::Float64
     weighted::Bool
     cache::Bool
+    prevalence::Union{Nothing, Float64}
 end
 
-Fluctuation(Ψ, initial_factors; tol=nothing, max_iter=1, ps_lowerbound=1e-8, weighted=false, cache=false) =
-    Fluctuation(Ψ, initial_factors, tol, max_iter, ps_lowerbound, weighted, cache)
+Fluctuation(Ψ, initial_factors; tol=nothing, max_iter=1, ps_lowerbound=1e-8, weighted=false, cache=false, prevalence=nothing) =
+    Fluctuation(Ψ, initial_factors, tol, max_iter, ps_lowerbound, weighted, cache, prevalence)
 
 one_dimensional_path(target_scitype::Type{T}) where T <: AbstractVector{<:MLJBase.Continuous} = LinearRegressor(fit_intercept=false, offsetcol = :offset)
 one_dimensional_path(target_scitype::Type{T}) where T <: AbstractVector{<:Finite} = LinearBinaryClassifier(fit_intercept=false, offsetcol = :offset)
@@ -55,6 +56,7 @@ function initialize_observed_cache(model, X, y)
         ps_lowerbound=model.ps_lowerbound,
         weighted_fluctuation=model.weighted
     )
+    w .*= get_weights_from_prevalence(model.prevalence, y)
     ŷ = MLJBase.predict(Q⁰, X)
     return Dict{Symbol, Any}(:H => H, :w => w, :ŷ => ŷ, :y => float(y))
 end
@@ -70,17 +72,20 @@ For each counterfactual generated by the indicator functions, we store:
 
 These are used to evaluate the gradient and estimate
 """
-function initialize_counterfactual_cache(model, X)
+function initialize_counterfactual_cache(model, X, y)
     Q⁰ = model.initial_factors.outcome_mean
     G⁰ = model.initial_factors.propensity_score
     Ψ = model.Ψ
 
-    counterfactual_cache = (predictions=[], signs=[], covariates=[], weights=[])
+    prevalence_weights = get_weights_from_prevalence(model.prevalence, y)
+    counterfactual_cache = (predictions=[], signs=[], covariates=[], weights=prevalence_weights)
     Ttemplate = selectcols(X, treatments(Ψ))
+    
     for (vals, sign) in indicator_fns(Ψ)
         T_ct = counterfactualTreatment(vals, Ttemplate)
         X_ct = DataFrame((;(Symbol(colname) => colname ∈ names(T_ct) ? T_ct[!, colname] : X[!, colname] for colname in names(X))...))
         
+        # w_ct is ununsed, passing prevalence_weights for downstream functions instead
         covariates_ct, w_ct = clever_covariate_and_weights(Ψ, 
             G⁰,
             X_ct; 
@@ -100,26 +105,23 @@ function initialize_counterfactual_cache(model, X)
             counterfactual_cache.covariates, 
             covariates_ct
         )
-        push!(
-            counterfactual_cache.weights, 
-            w_ct
-        )
     end
     return counterfactual_cache
 end
 
 function compute_counterfactual_aggregate!(counterfactual_cache, Q)
     ct_aggregate = zeros(length(first(counterfactual_cache.predictions)))
-    for (idx, (ct_ŷ, sign, ct_covariates)) in enumerate(zip(
+    for (idx, (ct_ŷ, sign, ct_covariates, weights)) in enumerate(zip(
             counterfactual_cache.predictions, 
             counterfactual_cache.signs, 
-            counterfactual_cache.covariates
+            counterfactual_cache.covariates,
[…]
