### 2025-07-16 9c8444a
make some abstraction in the MLConditionalDistributionEstimator
