### 2025-07-16 97a5468
add some methods for RieszRepresenter estimand
