### 2025-07-18 63d8a88
ready to work on rieszlearning integration
