+++ src/utils.jl
@@ -118,7 +118,7 @@ models = default_models(
 ```
 
 """
-default_models(;Q_binary=LinearBinaryClassifier(), Q_continuous=LinearRegressor(), G=LinearBinaryClassifier(), kwargs...) = Dict(
+default_models(;Q_binary=LinearBinaryClassifier(), Q_continuous=LinearRegressor(), G=LinearBinaryClassifier(), kwargs...) = Dict{Symbol, Any}(
     :Q_binary_default     => with_encoder(Q_binary),
     :Q_continuous_default => with_encoder(Q_continuous),
     :G_default            => with_encoder(G),
