### 2025-07-18 37c6151
push test manifest and update fluctuation for riesz
