+++ ext/RiesznetExt.jl
@@ -1,6 +1,0 @@
-module RiesznetExt
-
-using TMLE
-using RieszLearning
-
-end
