### 2025-07-18 8da0371
possibly fix cannot merge projects
