+++ Project.toml
@@ -47,7 +47,7 @@ CausalTables = "1.2.1"
 Combinatorics = "1.0.2"
 ComputationalResources = "0.3.2"
 DataFrames = "1.7.0"
-DifferentiationInterface = "0.6.43"
+DifferentiationInterface = "0.6, 0.7"
 Distributions = "0.25"
 GLM = "1.8.2"
 Graphs = "1.8"
