### 2025-07-20 b4d89a9
remove .components where needless and add interface for JointCondiitonalEstimate
