### 2025-07-20 1b8cadc
add more simplifications
