### 2025-07-21 f7c8ab4
fix some type signatures
