### 2025-07-22 ccf47ee
add some tests
