+++ test/counterfactual_mean_based/estimators_and_estimates.jl
@@ -11,6 +11,7 @@ using CategoricalArrays
 using LogExpFunctions
 using MLJBase
 using CSV
+using RieszLearning
 
 DATADIR = joinpath(pkgdir(TMLE), "test", "data")
 
@@ -138,6 +139,65 @@ end
         @test e.msg == TMLE.outcome_mean_fluctuation_fit_error_msg(Q)
     end
 end
+
+@testset "Test build_treatments_factor_estimator" begin
+    dataset = make_dataset()
+    Ψ = ATE(
+        outcome = :Y,
+        treatment_values = (T₁=(case=true, control=false),),
+        treatment_confounders = (T₁ = [:W],)
+    )
+    models = default_models()
+    # When the estimand is a JointConditionalDistribution
+    propensity_score = TMLE.JointConditionalDistribution(
+        TMLE.ConditionalDistribution(:T₁, [:W])
+    )
+    ## No sample split
+    treatments_factor_estimator = TMLE.build_treatments_factor_estimator(
+        propensity_score, 
+        models, 
+        dataset;
+        train_validation_indices=nothing
+    )
+    @test treatments_factor_estimator isa TMLE.JointConditionalDistributionEstimator
+    @test treatments_factor_estimator.cd_estimators[:T₁] isa TMLE.MLEstimator
+    ## Sample split
+    treatments_factor_estimator = TMLE.build_treatments_factor_estimator(
+        propensity_score, 
+        models, 
+        dataset;
+        train_validation_indices=[(1:50, 51:100), (51:100, 1:50)]
+    )
+    @test treatments_factor_estimator isa TMLE.JointConditionalDistributionEstimator
+    @test treatments_factor_estimator.cd_estimators[:T₁] isa TMLE.SampleSplitMLEstimator
+
+    # When the estimand in the RieszRepresenter
+    riesz_representer = TMLE.RieszRepresenter(Ψ)
+    models[:riesz_representer] = RieszLearning.RieszNetModel(
+        lux_model=RieszLearning.MLP([5, 10]),
+        hyper_parameters=(
+            batch_size=10,
+            nepochs=5
+        )
+    )
+    ## No sample split
+    treatments_factor_estimator = TMLE.build_treatments_factor_estimator(
+        riesz_representer, 
+        models, 
+        dataset; 
+        train_validation_indices=nothing
+    )
+    @test treatments_factor_estimator isa TMLE.MLEstimator
+    ## Sample split
+    treatments_factor_estimator = TMLE.build_treatments_factor_estimator(
+        riesz_representer, 
+        models, 
+        dataset; 
+        train_validation_indices=[(1:50, 51:100), (51:100, 1:50)]
+    )
+    @test treatments_factor_estimator isa TMLE.SampleSplitMLEstimator
+end
+
 
 @testset "Test structs are concrete types" begin
     for type in (Ose, Tmle, Plugin)
