### 2025-07-22 e09f947
add sample split riesz, need more tests
