### 2025-07-22 1f18d01
add some tests
