+++ test/counterfactual_mean_based/fluctuation.jl
@@ -6,6 +6,7 @@ using MLJModels
 using MLJBase
 using DataFrames
 using Distributions
+using StableRNGs
 
 @testset "Test Fluctuation with 1 Treatments" begin
     Ψ = ATE(
@@ -86,6 +87,115 @@ using Distributions
     fluctuation_mean = TMLE.expected_value(MLJBase.predict(unweighted_fluctuation, uw_machines, X))
     mse_fluct = sum((fluctuation_mean .- y).^2)
     @test mse_fluct < mse_initial 
+end
+
+@testset "Test Fluctuation with 1 Treatments: CV mode" begin
+    Ψ = ATE(
+        outcome=:Y,
+        treatment_confounders=(T=[:W₁, :W₂, :W₃],),
+        treatment_values=(T=(case="a", control="b"),),
+    )
+    dataset = DataFrame(
+        T = categorical(["a", "b", "c", "a", "a", "b", "a"]),
+        Y = [1., 2., 3, 4, 5, 6, 7],
+        W₁ = rand(7),
+        W₂ = rand(7),
+        W₃ = rand(7)
+    )
+    η = TMLE.CMRelevantFactors(
+        TMLE.ConditionalDistribution(:Y, [:T, :W₁, :W₂, :W₃]),
+        TMLE.JointConditionalDistribution(TMLE.ConditionalDistribution(:T, [:W₁, :W₂, :W₃]))
+    )
+    train_validation_indices = MLJBase.train_test_pairs(
+        CV(nfolds=3, shuffle=true, rng=StableRNG(123)), 
+        1:nrows(dataset), 
+        dataset
+    )
+    η̂ = TMLE.CMRelevantFactorsEstimator(
+        train_validation_indices,
+        Dict(
+            :Y => with_encoder(ConstantRegressor()), 
+            :T => ConstantClassifier()
+        )
+    )
+    η̂ₙ = η̂(η, dataset, verbosity = 0)
+    X, y = TMLE.get_mlj_inputs_and_target(η̂ₙ.outcome_mean.estimand, dataset)
+
+    mse_initial = sum((TMLE.expected_value(η̂ₙ.outcome_mean, X) .- y).^2)
+
+    weighted_fluctuation = TMLE.Fluctuation(Ψ, η̂ₙ; weighted=true)
+    ## First check the initialization of the counterfactual/observed caches
+    counterfactual_cache = TMLE.initialize_counterfactual_cache(weighted_fluctuation, X)
+    ### Predictions are independent of covariates here but computed per fold
+    expected_μ̂ = Vector{Float64}(undef, 7)
+    expected_weights_a = Vector{Float64}(undef, 7)
+    expected_weights_b = Vector{Float64}(undef, 7)
+    expected_weights_observed = Vector{Float64}(undef, 7)
+    for (index, (train_idx, val_idx)) in enumerate(train_validation_indices)
+        train_μ̂ = mean(dataset[train_idx, :Y])
+        train_pa = mean(dataset[train_idx, :T] .== "a")
+        train_pb = mean(dataset[train_idx, :T] .== "b")
+        train_pc = mean(dataset[train_idx, :T] .== "c")
+        expected_μ̂[val_idx] .= train_μ̂
+        expected_weights_a[val_idx] .= 1 / train_pa
+        expected_weights_b[val_idx] .= 1 / train_pb
+        expected_weights_observed[val_idx] = map(dataset[val_idx, :T]) do t
+            if t == "a"
+                1 / train_pa
+            elseif t == "b"
+                1 / train_pb
+            else 
+                1e8 ## done by ConstantClassifier
+            end
+        end
+    end
+    @test mean.(counterfactual_cache.predictions[1]) == expected_μ̂
[…]
