+++ src/counterfactual_mean_based/estimates.jl
@@ -22,7 +22,7 @@ string_repr(estimate::MLCMRelevantFactors) = string(
 )
 
 #####################################################################
-###                       FoldsMLCMRelevantFactors                     ###
+###                  FoldsMLCMRelevantFactors                     ###
 #####################################################################
 
 struct FoldsMLCMRelevantFactors <: Estimate
