### 2025-07-22 7d087d4
add riesz test to test suite
