### 2025-07-22 eab438e
remove manifest and add source section
