### 2025-07-23 e52dda9
add some possible transitions to CTMLE with Riesz
