+++ src/counterfactual_mean_based/fluctuation.jl
@@ -6,11 +6,11 @@ mutable struct Fluctuation <: MLJBase.Supervised
     ps_lowerbound::Float64
     weighted::Bool
     cache::Bool
-    prevalence::Union{Nothing, Float64}
+    prevalence_weights::Union{Nothing, Vector{Float64}}
 end
 
-Fluctuation(Ψ, initial_factors; tol=nothing, max_iter=1, ps_lowerbound=1e-8, weighted=false, cache=false, prevalence=nothing) =
-    Fluctuation(Ψ, initial_factors, tol, max_iter, ps_lowerbound, weighted, cache, prevalence)
+Fluctuation(Ψ, initial_factors; tol=nothing, max_iter=1, ps_lowerbound=1e-8, weighted=false, cache=false, prevalence_weights=nothing) =
+    Fluctuation(Ψ, initial_factors, tol, max_iter, ps_lowerbound, weighted, cache, prevalence_weights)
 
 one_dimensional_path(target_scitype::Type{T}) where T <: AbstractVector{<:MLJBase.Continuous} = LinearRegressor(fit_intercept=false, offsetcol = :offset)
 one_dimensional_path(target_scitype::Type{T}) where T <: AbstractVector{<:Finite} = LinearBinaryClassifier(fit_intercept=false, offsetcol = :offset)
@@ -56,7 +56,7 @@ function initialize_observed_cache(model, X, y)
         ps_lowerbound=model.ps_lowerbound,
         weighted_fluctuation=model.weighted
     )
-    w .*= get_weights_from_prevalence(model.prevalence, y)
+    w .*= model.prevalence_weights
     ŷ = MLJBase.predict(Q⁰, X)
     return Dict{Symbol, Any}(:H => H, :w => w, :ŷ => ŷ, :y => float(y))
 end
@@ -72,20 +72,18 @@ For each counterfactual generated by the indicator functions, we store:
 
 These are used to evaluate the gradient and estimate
 """
-function initialize_counterfactual_cache(model, X, y)
+function initialize_counterfactual_cache(model, X)
     Q⁰ = model.initial_factors.outcome_mean
     G⁰ = model.initial_factors.propensity_score
     Ψ = model.Ψ
 
-    prevalence_weights = get_weights_from_prevalence(model.prevalence, y)
-    counterfactual_cache = (predictions=[], signs=[], covariates=[], weights=prevalence_weights)
+    counterfactual_cache = (predictions=[], signs=[], covariates=[], weights = [], prevalence_weights=model.prevalence_weights)
     Ttemplate = selectcols(X, treatments(Ψ))
     
     for (vals, sign) in indicator_fns(Ψ)
         T_ct = counterfactualTreatment(vals, Ttemplate)
         X_ct = DataFrame((;(Symbol(colname) => colname ∈ names(T_ct) ? T_ct[!, colname] : X[!, colname] for colname in names(X))...))
         
-        # w_ct is ununsed, passing prevalence_weights for downstream functions instead
         covariates_ct, w_ct = clever_covariate_and_weights(Ψ, 
             G⁰,
             X_ct; 
@@ -105,23 +103,28 @@ function initialize_counterfactual_cache(model, X, y)
             counterfactual_cache.covariates, 
             covariates_ct
         )
+        push!(
+            counterfactual_cache.weights, 
+            w_ct
+        )
     end
     return counterfactual_cache
 end
 
 function compute_counterfactual_aggregate!(counterfactual_cache, Q)
     ct_aggregate = zeros(length(first(counterfactual_cache.predictions)))
-    for (idx, (ct_ŷ, sign, ct_covariates, weights)) in enumerate(zip(
+    for (idx, (ct_ŷ, sign, ct_covariates, weights, prevalence_weights)) in enumerate(zip(
[…]
