### 2025-07-30 db20019
Removed test for non-existent util; Revert Project.toml
