### 2025-07-30 f96b2ff
Small fix to fluctuation behvaior in dealing with CCW
