### 2025-07-31 160eb2f
move todolist away from readme
