### 2025-07-31 2a1529f
add DifferentiationInterfaceExt
