+++ Project.toml
@@ -9,7 +9,6 @@ CategoricalArrays = "324d7699-5711-5eae-9e2f-1d82baa6b597"
 Combinatorics = "861a8166-3701-5b0c-9a16-15d98fcdc6aa"
 ComputationalResources = "ed09eef8-17a6-5b46-8889-db040fac31e3"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
-DifferentiationInterface = "a0c0ee7d-e4b9-4e03-894e-1c5f64a51d63"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 GLM = "38e38edf-8417-5370-95a0-9cbb8c7f171a"
 Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
@@ -26,14 +25,15 @@ Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 StatisticalMeasures = "a19d573c-0a75-4610-95b3-7071388c7541"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
-Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [weakdeps]
+DifferentiationInterface = "a0c0ee7d-e4b9-4e03-894e-1c5f64a51d63"
 CausalTables = "6af48e0c-efc2-4bf7-a92f-a553ccf79fd6"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
 YAML = "ddb6d928-2868-570f-bddf-ab3f9cf99eb6"
 
 [extensions]
+DifferentiationInterfaceExt = "DifferentiationInterface"
 CausalTablesExt = "CausalTables"
 JSONExt = "JSON"
 YAMLExt = "YAML"
@@ -62,7 +62,6 @@ Printf = "1.11.0,1.10.9"
 StatisticalMeasures = "0.2"
 Tables = "1.6"
 YAML = "0.4.9"
-Zygote = "0.7"
 julia = "1.10, 1"
 
 [extras]
