### 2025-07-31 TARGENE/TMLE.jl#139 @olivierlabayle
Extension for SCM
It would be useful to trim the package further by having an extension to deal with Structural Causal Models. The Julia ecosystem does not have a strong candidate for this at the moment.

- CausalInference.jl: works with simple graphs with no annotations
- CausalTables.jl: requires a dgp and further metadata.

All we need isa graph with annotations like provided by MetaGraphsNext.jl on which we could build an extension. I am just considering it could result in type piracies.
