### 2025-07-31 213ceec
update docs and make two non documented methods private
