+++ docs/Project.toml
@@ -4,6 +4,7 @@ CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
 CategoricalArrays = "324d7699-5711-5eae-9e2f-1d82baa6b597"
 CausalTables = "6af48e0c-efc2-4bf7-a92f-a553ccf79fd6"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
+DifferentiationInterface = "a0c0ee7d-e4b9-4e03-894e-1c5f64a51d63"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 GLM = "38e38edf-8417-5370-95a0-9cbb8c7f171a"
@@ -25,6 +26,7 @@ TMLE = "8afdd2fb-6e73-43df-8b62-b1650cd9c8cf"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 YAML = "ddb6d928-2868-570f-bddf-ab3f9cf99eb6"
+Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [compat]
 CSV = "0.10.11"
@@ -39,3 +41,5 @@ MLJModels = "0.17"
 MLJXGBoostInterface = "0.3.8"
 NearestNeighborModels = "0.2"
 StableRNGs = "1.0"
+DifferentiationInterface = "0.7"
+Zygote = "0.7"
