### 2025-07-31 9478285
fix warnings on 3 point test
