### 2025-07-31 155d417
update categorical array compat
