### 2025-08-01 #8446 @olivierlabayle
Just letting you know I've also made a Julia extension for AbstractDifferentiation and Zygote which should further reduce the size of the package. I've opened an issue to potentially also add an extension for SCMs but there is not strong candidate for that yet I think: https://github.com/TARGENE/TMLE.jl/issues/139.
