### 2025-08-01 c731d56
Explicitly passes prevalence weights as actual weights or nothing type (no more ones() vectors); Error is thrown when a model is specifed that does not accept weights; Removed unnessesary dependency
