+++ Project.toml
@@ -17,7 +17,6 @@ HypothesisTests = "09f84164-cd44-5f33-b23f-e6b0d136a0d5"
 LogExpFunctions = "2ab3a3ac-af41-5b50-aa03-7779005ae688"
 MLJBase = "a7f614a8-145f-11e9-1d2a-a57a1082229d"
 MLJGLMInterface = "caf8df21-4939-456d-ac9c-5fefbfb04c0c"
-MLJLinearModels = "6ee0df7b-362f-4a72-a706-9e79364fb692"
 MLJModels = "d491faf4-2d78-11e9-2867-c94bc002c0b7"
 MetaGraphsNext = "fa8bd995-216d-47f1-8a91-f3b68fbeb377"
 Missings = "e1d29d7a-bbdc-5cf2-9ac0-f12de2c33e28"
@@ -57,7 +56,6 @@ JSON = "0.21.4"
 LogExpFunctions = "0.3"
 MLJBase = "1"
 MLJGLMInterface = "0.3.4"
-MLJLinearModels = "0.10.1"
 MLJModels = "0.15, 0.16, 0.17"
 MetaGraphsNext = "0.7"
 Missings = "1.0"

[…]
