### 2025-08-03 88eeec3
expectation weighting fixed; removal of the mariginal classes; easier weight subsetting
