### 2025-08-05 bec3324
Testing for model weights
