### 2025-08-05 be3e1b3
Remove unnessesary weighting from the compuation of the ctf_agg
