### 2025-08-06 d763f7a
Ensure outcome is binary in CCW config
