+++ src/utils.jl
@@ -208,3 +208,26 @@ outcome_mean_fluctuation_fit_error_msg(factor) = string(
 Base.showerror(io::IO, e::FitFailedError) = print(io, e.msg)
 
 with_encoder(model; encoder=ContinuousEncoder(drop_last=true, one_hot_ordered_factors = false)) = Pipeline(encoder,  model)
+
+function is_binary_column(df::DataFrame, col::Symbol)
+    vals = unique(skipmissing(df[!, col]))
+    return length(vals) == 2 && all(x -> x in (0, 1) || x in (true, false), vals)
+end
+
+"""
+    ccw_check(prevalence::Union{Nothing, Float64}, dataset, relevant_factors)
+
+Check if the dataset is suitable for prevalence correction (CCW-TMLE) throws an error if the outcome column is not binary when prevalence is provided.
+If the dataset is suitable, it returns the dataset with missing values dropped from the outcome column.
+
+"""
+function ccw_check(prevalence::Union{Nothing, Float64}, dataset, relevant_factors)
+    if !isnothing(prevalence)
+        dataset = dropmissing(dataset, relevant_factors.outcome_mean.outcome)
+        is_binary_column(dataset, relevant_factors.outcome_mean.outcome) || 
+            throw(ArgumentError("Outcome column must be binary for prevalence correction (CCW-TMLE)."))
+        return dataset
+    else
+        return dataset
+    end
+end
