+++ src/counterfactual_mean_based/estimators.jl
@@ -108,6 +108,8 @@ function (tmle::Tmle)(Ψ::StatisticalCMCompositeEstimand, dataset; cache=Dict(),
     # Initial fit of the SCM's relevant factors
     relevant_factors = get_relevant_factors(Ψ, collaborative_strategy=tmle.collaborative_strategy)
     nomissing_dataset = nomissing(dataset, variables(relevant_factors))
+    # If prevalence is provided, we must ensure that the dataset has no missing values in the outcome column
+    dataset = ccw_check(tmle.prevalence, dataset, relevant_factors)
     initial_factors_dataset = choose_initial_dataset(dataset, nomissing_dataset, train_validation_indices)
     prevalence_weights = get_weights_from_prevalence(tmle.prevalence, initial_factors_dataset[!, relevant_factors.outcome_mean.outcome])
 
