### 2025-08-06 cb3af5d
updating weighted gradient; addressing latest comments; fixing tests
