### 2025-08-08 6445c79
try up manifest
