+++ .github/workflows/CI.yml
@@ -35,6 +35,7 @@ jobs:
             ${{ runner.os }}-test-
             ${{ runner.os }}-
       - uses: julia-actions/julia-buildpkg@v1
+      - run: julia --project=test -e 'using Pkg; Pkg.instantiate()' # Necessary for now, otherwise unregistered package RieszLearning will trigger an error 
       - uses: julia-actions/julia-runtest@v1
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v1
