### 2025-08-08 9c334a4
try add resolve as well
