+++ Project.toml
@@ -67,6 +67,7 @@ TableOperations = "1.2"
 Tables = "1.6"
 YAML = "0.4.9"
 Zygote = "0.7"
+Statistics = "1.10, 1"
 julia = "1.10, 1"
 
 [extras]
