### 2025-08-08 b1a05a8
added check_inputs(); updated test robustness; increased code readability
