+++ test/runtests.jl
@@ -28,11 +28,9 @@ TEST_DIR = joinpath(pkgdir(TMLE), "test")
     @test include(joinpath(TEST_DIR, "counterfactual_mean_based/covariate_based_strategies.jl"))
     
     # Test Extensions
-    if VERSION >= v"1.9"
-        @test include(joinpath(TEST_DIR, "configuration.jl"))
-        @test include(joinpath(TEST_DIR, "causaltables_interface.jl"))
-        @test include(joinpath(TEST_DIR, "riesz_learning.jl"))
-    end
+    @test include(joinpath(TEST_DIR, "configuration.jl"))
+    @test include(joinpath(TEST_DIR, "causaltables_interface.jl"))
+    @test include(joinpath(TEST_DIR, "riesz_learning.jl"))
 
     # Test Experimental
     @test include(joinpath(TEST_DIR, "estimand_ordering.jl"))
