+++ Project.toml
@@ -42,7 +42,7 @@ YAMLExt = "YAML"
 
 [compat]
 AutoHashEquals = "2.1.0"
-CategoricalArrays = "0.10, 1"
+CategoricalArrays = "0.10"
 CausalTables = "1.2.1"
 Combinatorics = "1.0.2"
 ComputationalResources = "0.3.2"
@@ -63,11 +63,11 @@ OrderedCollections = "1.6.3"
 Printf = "1.11.0,1.10.9"
 SplitApplyCombine = "1.2.2"
 StatisticalMeasures = "0.2"
+Statistics = "1.10, 1"
 TableOperations = "1.2"
 Tables = "1.6"
 YAML = "0.4.9"
 Zygote = "0.7"
-Statistics = "1.10, 1"
 julia = "1.10, 1"
 
 [extras]
