### 2025-08-08 065dda9
testing CCW_IC; simplify check_inputs()
