### 2025-08-08 cebfac4
add extra test for riesz learning
