### 2025-08-08 92df9a4
reduce_gradient();slight_fix to ccw_check
