### 2025-08-08 98a0b1f
provide test for the expection of the IC
