### 2025-08-09 #8446 @nhejazi
sorry for the long delay @olivierlabayle @jbytecode, it took some time to be able to work through this. the package and documentation are very nicely put together, so i'm providing some comments to document my experience installing it and taking it for a test run and, separately, comments on the paper as currently written. some specific comments below

### the package

to try the package out, i cloned the repo from branch `ol_dev_doc`, updated requirements with `] up`, and ran the test suite with `] test` on Julia v1.11.6, the latter of which showed passing tests:
```julia
153.727880 seconds (1.52 G allocations: 85.220 GiB, 3.05% gc time, 110 lock conflicts, 76.26% compilation time: 3% of which was recompilation)
     Testing TMLE tests passed
```
but some of the tests produced seemingly trivial warnings, e.g.,
```julia
┌ Warning: `TMLEE(args...; kwargs...)` is deprecated, use `Tmle(args...; kwargs...)` instead.
│   caller = macro expansion at Test.jl:677 [inlined]
```
which seems to just point out that some refactoring was done but not propagated to the tests, and some issues related to MLJ, e.g.,
```julia
┌ Error: Problem fitting the machine machine(LogisticClassifier(lambda = 2.220446049250313e-16, …), …).
└ @ MLJBase ~/.julia/packages/MLJBase/v5sA7/src/machines.jl:694
[ Info: Running type checks...
┌ Warning: The number and/or types of data arguments do not match what the specified model
│ supports. Suppress this type check by specifying `scitype_check_level=0`.
```
which could be suppressed, i guess, as MLJ seems to suggest. After this, i ran the [quick start example](https://targene.github.io/TMLE.jl/previews/PR135/#Quick-Start) in the documentation webpage and, reassuringly, was able to reproduce the results exactly:
```julia
julia> result
One sample t-test
-----------------
Population details:
    parameter of interest:   Mean
    value under h_0:         0
    point estimate:          2.49315
    95% confidence interval: (2.434, 2.552)

Test summary:
    outcome with 95% confidence: reject h_0
    two-sided p-value:           <1e-93

Details:
    number of observations:   100
    t-statistic:              83.82654382411157
    degrees of freedom:       99
    empirical standard error: 0.029741766265683135
```
from a user-facing perspective, a few suggestions on the above output would be (1) to consider truncating estimates in the "Details" section to some reasonable degree (e.g., 4 decimal places)  and (2) to also truncate the p-value reported more aggressively, e.g., 1e-10 as a minimum. In terms of the reported output, the object `result` contains the estimated influence function in the slot `result.IC`, and i would recommend reporting the mean of this as a diagnostic, as the IC should be, by theory, mean zero, and empirical deviations from this can provide a first-order assessment of estimator quality (since the targeting step essentially approximately solves the estimating equation $P_n IC(P_0)(O) \approx 0$.

since the package's tests pass and the quick start example reproduced the output on the documentation webpage, i've checked off the items in the review checklist related to the software package. overall, i think the package should be a welcome contribution to the julia ecosystem and provides important, useful functionality to support causal machine learning; moreover, the documentation webpage is informative and the availability of several estimands (and extensibility to continue adding more) should make the package widely usable.

### the paper

while JOSS does not provide more than minimal specifications for submitted papers and noting that my intention is not to be prescriptive as to how the paper should be written, my suggestions center on substantially rewriting certain aspects of the draft as elaborated upon below.

- the "Summary" section of the paper explains TMLE as a procedure seemingly in a class of its own, which is somewhat misleading, especially given that the package also implements the (asymptotically equivalent) one-step estimator. it would be more useful---and correct---to contextualize TMLE within the broader landscape of causal/de-biased machine learning, inclusive of the reliance on semi-parametric estimation and causal inference to guide choice of statistical estimands that all such procedures in this area share
- the "Statement of Need" section mentions a handful of packages (`tmle`, `tmle3`, and `ltmle`, specifically) without referencing many additional packages that have been written and are widely used by those interested in implementing TMLE. it would help to provide some context, e.g., briefly stating what `tmle` and `ltmle` do and how they differ as well as how `tmle3` and the `TLverse` ecosystem of which it is a part are distinct from earlier implementations. this section then goes on to provide a bulleted list of estimands and estimators available in `TMLE.jl` with very limited description of these. while it is not necessary to reproduce the documentation in the paper, it's not clear to me what i would understand from this without a fairly significant existing background in this area. i'd strongly suggest rewriting this in a more explanatory form, without the bulleted lists and focusing on getting key ideas across.
- the "Mentions" section again simply lists a few use-cases. it would help the reader, i think, to instead describe in a couple of sentences what these applied science problems were as well as (1) how the package helped to solve underlying statistical learning problems and (2) how being brought to bear on these examples improved the package's design if at all. perhaps the section could be renamed "applied science use-cases" or similar.
