### 2025-08-11 #8446 @salbalkus
Hi @olivierlabayle, building off of @nhejazi 's comments, I also have a few of my own:

**Paper comments**
- Agree with Nima -- "Summary" should also quickly describe the field of "causal machine learning" in which TMLE lies. I think this will provide broader context for readers with different backgrounds and help people who don't know TMLE but are generally familiar with causal inference understand how this package is situated in the literature.
- In light of the above, I think the paper should mention some broader software packages for causal machine learning in general, such as [EconML](https://github.com/py-why/EconML), [CausalML](https://causalml.readthedocs.io/en/latest/about.html), or [DoubleML](https://docs.doubleml.org/stable/index.html) which also use machine learning to estimate causal effects.
- Also similar to Nima's comment: I think the sub-bullets in the "Statement of Need" section could be converted into a bit more of a narrative for each bullet point,  providing a bit more context for why these features are included.

**Documentation comments**
- Explanations are very thorough -- well done on this point.
- The one main critique I have is that the "Overview" reads more like an introduction to a paper about TMLE rather than a quick overview of what the package *does*. I do like the writing as background, but maybe it would be better to put it into a different / later subsection called something like "Why TMLE?" and have the start of the landing page describe what the method does (i.e. "**use targeted / doubly robust machine learning in Julia to estimate causal estimands such as the ATE, counterfactual mean**, *other estimands* etc). It's important to have a concise description to "sell" the package and state clearly its functionality to people unfamiliar with what TMLE is right at the start.
- I like the Quick Start example. Though, it might be a good idea to include the analysis of a "real" dataset prominently as an example somewhere, just to emphasize how TMLE is meant to be used in a practical setting. This is especially true for the "Walkthrough" where you have a very large simulated data set that might be confusing to beginners. I've found a dataset like Boston Housing in a package like [MLDatasets.jl](https://juliaml.github.io/MLDatasets.jl/stable/datasets/misc/#Miscellaneuous-Datasets) can be quite useful to replace simulations in certain settings like this.

Overall, nice work. Once these issues are addressed I will complete my checklist!
