### 2025-08-13 b5bdf0a
add custom printing of estimates
