+++ paper/paper.md
@@ -39,44 +39,29 @@ bibliography: paper.bib
 
 # Summary
 
-Causal inference is essential for understanding the effect of interventions in real-world settings, such as determining whether a treatment improves health outcomes or whether a gene variant contributes to disease risk. Traditional statistical methods, such as linear regression or propensity score matching, often rely on strong modeling assumptions and may fail to provide valid inference when these assumptions are violated—particularly in the presence of high-dimensional data or model misspecification.
+Causal inference is essential for understanding the effects of interventions in real-world settings, such as determining whether a treatment improves health outcomes or whether a genetic variant contributes to disease risk. Traditional approaches often begin by positing a specific parametric model, such as a linear-Gaussian or logistic regression model, and then estimating its parameters using efficient likelihood-based methods. This strategy has two main drawbacks. First, the quantity being estimated is often dictated by the parametric form of the model rather than the scientific question of interest (e.g., additive effects under a linear model, odds ratios under a logistic model). Second, when the model is misspecified, particularly in high-dimensional settings with complex interactions, parametric estimators can be severely biased.
 
-TMLE.jl is a Julia package that implements Targeted Maximum Likelihood Estimation (TMLE) [@van2011targeted;@van2018targeted], a general framework for causal effect estimation that combines machine learning with principles from semiparametric statistics. TMLE provides doubly robust, efficient, and flexible estimation of causal parameters in observational and experimental studies.
+Over the past two decades, new approaches have emerged that combine causal inference, machine learning (ML), and semiparametric theory to address these limitations. These methods begin with the explicit definition of a target parameter, often a causal estimand such as the average treatment effect, derived from a causal model. They then replace restrictive parametric models with flexible ML algorithms (e.g., gradient boosting, neural networks) to estimate nuisance components such as the outcome regression and treatment mechanism (called nuisance functions). However, since most ML methods are optimized for prediction rather than unbiased estimation, ML-based estimates are still biased when plugged directly into the parameter formula. To remove this bias, modern approaches use a debiasing step based on a key mathematical object, the efficient influence function (EIF). The result is an estimator that is asymptotically unbiased, efficient, and valid under much weaker assumptions than traditional parametric models.
 
-The goal of TMLE.jl is to provide an accessible implementation of this methodology within the Julia ecosystem. It enables researchers to estimate average treatment effects and other causal parameters while leveraging modern machine learning algorithms to flexibly model nuisance components, such as the outcome regression and treatment mechanism.
[…]
