### 2025-08-13 ee4e6d6
fix import and add a introductory sentence to the landing page
