### 2025-08-13 78c7583
slight update of home text
