### 2025-08-13 #8446 @salbalkus
Thanks for the update @olivierlabayle! I think the paper and documentation are looking much better now. The only concern I have at the moment is that the "Summary" section is very long (about half the current paper) -- in other JOSS papers I've seen it's generally a 1-paragraph overview of the package's functionality. I think it might be worth moving exactly what you've written for the Summary to make it the first part of "Statement of Need" and then writing a quick few sentences to summarize the package. Even the second paragraph of the TMLE.jl README on GitHub would make a good summary for the package as a whole, I think. But this is just a narrative / stylistic preference, if other reviewers or the editor have different opinions, happy to defer to them.
