### 2025-08-14 a1a03dc
make new summary section and push old summary to background
