+++ paper/paper.md
@@ -39,6 +39,10 @@ bibliography: paper.bib
 
 # Summary
 
+TMLE.jl is a Julia package implementing targeted minimum loss based estimation (TMLE), a general framework for causal effect estimation that unites modern machine learning with the theoretical guarantees of semiparametric statistics. TMLE yields doubly robust and semiparametrically efficient estimators, meaning it remains consistent if either the outcome model or the treatment assignment model is correctly specified, and it achieves the smallest possible asymptotic variance under standard regularity conditions. The package integrates with the broader Julia machine learning ecosystem and can be used in both observational and experimental settings. It is particularly well-suited for high-dimensional problems where robust inference is essential.
+
+# Background
+
 Causal inference is essential for understanding the effects of interventions in real-world settings, such as determining whether a treatment improves health outcomes or whether a genetic variant contributes to disease risk. Traditional approaches often begin by positing a specific parametric model, such as a linear-Gaussian or logistic regression model, and then estimating its parameters using efficient likelihood-based methods. This strategy has two main drawbacks. First, the quantity being estimated is often dictated by the parametric form of the model rather than the scientific question of interest (e.g., additive effects under a linear model, odds ratios under a logistic model). Second, when the model is misspecified, particularly in high-dimensional settings with complex interactions, parametric estimators can be severely biased.
 
 Over the past two decades, new approaches have emerged that combine causal inference, machine learning (ML), and semiparametric theory to address these limitations. These methods begin with the explicit definition of a target parameter, often a causal estimand such as the average treatment effect, derived from a causal model. They then replace restrictive parametric models with flexible ML algorithms (e.g., gradient boosting, neural networks) to estimate nuisance components such as the outcome regression and treatment mechanism (called nuisance functions). However, since most ML methods are optimized for prediction rather than unbiased estimation, ML-based estimates are still biased when plugged directly into the parameter formula. To remove this bias, modern approaches use a debiasing step based on a key mathematical object, the efficient influence function (EIF). The result is an estimator that is asymptotically unbiased, efficient, and valid under much weaker assumptions than traditional parametric models.
