### 2025-08-14 #8446 @olivierlabayle
> 👉📄 [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.08446/joss.08446/10.21105.joss.08446.pdf) 📄 [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.08446/joss.08446/10.21105.joss.08446.pdf) 📄 👈

As suggested, I have made a very short summary section and pushed the old summary to a background section. I think this should address most if not all comments.
