### 2025-08-14 #8446 @nhejazi
thank you @olivierlabayle for the explanations of the unit tests and the updates to the package interface. the updated paper reads quite nicely and now provides more thorough coverage of the background and details on the package--excellent. as i was reading the paper, i noticed a typo: "isntance" instead of "instance" on line 82 of page 2; it may be worth re-reading closely again in case there are any more. otherwise, i think the package and paper are in great shape, and i've checked off the remaining items in my reviewer checklist, @jbytecode.
