### 2025-08-15 5cb735d
add citation entry for Julia in bibtex
