### 2025-08-15 559ab09
add citation for Julia in manuscript
