### 2025-08-15 8f4da0c
mark release number as breaking since some deps have been converted to extensions
