+++ paper/paper.bib
@@ -52,12 +52,12 @@
   number = {106}, 
   pages = {7580}, 
   author = {Salvador V. Balkus and Nima S. Hejazi}, 
-  title = {CausalTables.jl: Simulating and storing data for statistical causal inference in Julia}, 
+  title = {CausalTables.jl: Simulating and storing data for statistical causal inference in {J}ulia}, 
   journal = {Journal of Open Source Software} 
 }
 
 @misc{schäfer2022abstractdifferentiationjlbackendagnosticdifferentiableprogramming,
-      title={AbstractDifferentiation.jl: Backend-Agnostic Differentiable Programming in Julia}, 
+      title={AbstractDifferentiation.jl: Backend-Agnostic Differentiable Programming in {J}ulia}, 
       author={Frank Schäfer and Mohamed Tarek and Lyndon White and Chris Rackauckas},
       year={2022},
       eprint={2109.12449},
@@ -108,14 +108,14 @@
 }
 
 @article{blaom2020mlj,
-  title={MLJ: A Julia package for composable machine learning},
+  title={MLJ: A {J}ulia package for composable machine learning},
   author={Blaom, Anthony D and Kiraly, Franz and Lienart, Thibaut and Simillides, Yiannis and Arenas, Diego and Vollmer, Sebastian J},
   journal={arXiv preprint arXiv:2007.12285},
   year={2020}
 }
 
 @article{bouchet2023dataframes,
-  title={Dataframes. jl: Flexible and fast tabular data in julia},
+  title={{Dataframes.jl}: Flexible and fast tabular data in {J}ulia},
   author={Bouchet-Valat, Milan and Kami{\'n}ski, Bogumi{\l}},
   journal={Journal of Statistical Software},
   volume={107},
@@ -191,7 +191,7 @@
 }
 
 @software{DoubleML,
-  title    = {{DoubleML} -- Double Machine Learning in Python},
+  title    = {{DoubleML} -- {D}ouble Machine Learning in {P}ython},
   author   = {Bach, Philipp and Chernozhukov, Victor and Klaassen, Sven and Kurz, Malte S. and Spindler, Martin},
   year     = {2024},
   version  = {latest},
