### 2025-08-15 #8446 @jbytecode
@olivierlabayle - I've just sent a pull request (https://github.com/TARGENE/TMLE.jl/pull/141)

This pull request includes some minor fixes in both the manuscript and the bibliography.

Please review the changes. If you're agree with them, please merge.

Thank you in advance.
