+++ paper/paper.bib
@@ -2,7 +2,8 @@
   title={Targeted learning in data science},
   author={Van der Laan, Mark J and Rose, Sherri},
   year={2018},
-  publisher={Springer}
+  publisher={Springer},
+  doi={https://doi.org/10.1007/978-3-319-65304-4}
 }
 
 @article{smith2023application,
@@ -12,7 +13,8 @@
   volume={86},
   pages={34--48},
   year={2023},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.annepidem.2023.06.004}
 }
 
 @book{van2011targeted,
@@ -20,7 +22,8 @@
   author={Van der Laan, Mark J and Rose, Sherri and others},
   volume={4},
   year={2011},
-  publisher={Springer}
+  publisher={Springer},
+  doi={https://doi.org/10.1007/978-1-4419-9782-1}
 }
 
 @article{gruber2010application,
@@ -30,18 +33,19 @@
   volume={6},
   number={1},
   year={2010},
-  publisher={De Gruyter}
+  publisher={De Gruyter},
+  doi={10.2202/1557-4679.1182}
 }
 
-@misc{dalle2025commoninterfaceautomaticdifferentiation,
-      title={A Common Interface for Automatic Differentiation}, 
-      author={Guillaume Dalle and Adrian Hill},
-      year={2025},
-      eprint={2505.05542},
-      archivePrefix={arXiv},
-      primaryClass={cs.MS},
-      url={https://arxiv.org/abs/2505.05542}, 
+@software{dalleDifferentiationInterface2025,
+      author={Dalle, Guillaume and Hill, Adrian},
+      title={Differentiation{I}nterface.jl},
+      year={2024},
+      publisher={Zenodo},
+      doi={10.5281/zenodo.11092033},
+      url={https://doi.org/10.5281/zenodo.11092033},
 }
+
 
 @article{Balkus2025, 
   doi = {10.21105/joss.07580}, 
@@ -56,16 +60,6 @@
   journal = {Journal of Open Source Software} 
 }
 
-@misc{schäfer2022abstractdifferentiationjlbackendagnosticdifferentiableprogramming,
-      title={AbstractDifferentiation.jl: Backend-Agnostic Differentiable Programming in Julia}, 
-      author={Frank Schäfer and Mohamed Tarek and Lyndon White and Chris Rackauckas},
-      year={2022},
-      eprint={2109.12449},
-      archivePrefix={arXiv},
-      primaryClass={cs.MS},
-      url={https://arxiv.org/abs/2109.12449}, 
-}
-
 @article{lendle2017ltmle,
   title = {{ltmle}: An {R} Package Implementing Targeted Minimum Loss-Based Estimation for Longitudinal Data},
   author = {Samuel D. Lendle and Joshua Schwab and Maya L. Petersen and Mark J. {van der Laan}},
@@ -78,6 +72,7 @@
 }
 
 @article{labayle2025semi,
+  doi={https://doi.org/10.48550/arXiv.2505.14675},
   title={Semi-parametric efficient estimation of small genetic effects in large-scale population cohorts},
   author={Labayle, Olivier and Roskams-Hieter, Breeshey and Slaughter, Joshua and Tetley-Campbell, Kelsey and van der Laan, Mark J and Ponting, Chris P and Beentjes, Sjoerd Viktor and Khamseh, Ava},
   journal={arXiv preprint arXiv:2505.14675},
@@ -85,16 +80,16 @@
 }
 
 @article{tmleR,
-    title = {{tmle}: An {R} Package for Targeted Maximum Likelihood
-      Estimation},
[…]
