### 2025-08-15 2793ab3
remove doi prefixes
