+++ paper/paper.bib
@@ -3,7 +3,7 @@
   author={Van der Laan, Mark J and Rose, Sherri},
   year={2018},
   publisher={Springer},
-  doi={https://doi.org/10.1007/978-3-319-65304-4}
+  doi={10.1007/978-3-319-65304-4}
 }
 
 @article{smith2023application,
@@ -23,7 +23,7 @@
   volume={4},
   year={2011},
   publisher={Springer},
-  doi={https://doi.org/10.1007/978-1-4419-9782-1}
+  doi={10.1007/978-1-4419-9782-1}
 }
 
 @article{gruber2010application,
@@ -43,13 +43,13 @@
       year={2024},
       publisher={Zenodo},
       doi={10.5281/zenodo.11092033},
-      url={https://doi.org/10.5281/zenodo.11092033},
+      url={10.5281/zenodo.11092033},
 }
 
 
 @article{Balkus2025, 
   doi = {10.21105/joss.07580}, 
-  url = {https://doi.org/10.21105/joss.07580}, 
+  url = {10.21105/joss.07580}, 
   year = {2025}, 
   publisher = {The Open Journal}, 
   volume = {10}, 
@@ -72,7 +72,7 @@
 }
 
 @article{labayle2025semi,
-  doi={https://doi.org/10.48550/arXiv.2505.14675},
+  doi={10.48550/arXiv.2505.14675},
   title={Semi-parametric efficient estimation of small genetic effects in large-scale population cohorts},
   author={Labayle, Olivier and Roskams-Hieter, Breeshey and Slaughter, Joshua and Tetley-Campbell, Kelsey and van der Laan, Mark J and Ponting, Chris P and Beentjes, Sjoerd Viktor and Khamseh, Ava},
   journal={arXiv preprint arXiv:2505.14675},
@@ -98,12 +98,12 @@
   year = {2021},
   howpublished = {\url{https://github.com/tlverse/tmle3}},
   note = {{R} package version 0.2.0},
-  url = {https://doi.org/10.5281/zenodo.4603358},
+  url = {10.5281/zenodo.4603358},
   doi = {10.5281/zenodo.4603358}
 }
 
 @article{blaom2020mlj,
-  doi={https://doi.org/10.21105/joss.02704},
+  doi={10.21105/joss.02704},
   title={MLJ: A {J}ulia package for composable machine learning},
   author={Blaom, Anthony D and Kiraly, Franz and Lienart, Thibaut and Simillides, Yiannis and Arenas, Diego and Vollmer, Sebastian J},
   journal={arXiv preprint arXiv:2007.12285},
@@ -127,7 +127,7 @@
   pages={207--236},
   year={2024},
   publisher={Chapman and Hall/CRC},
-  doi={https://doi.org/10.48550/arXiv.2203.06469}
+  doi={10.48550/arXiv.2203.06469}
 }
 
 @article{pfanzagl1985contributions,
@@ -139,7 +139,7 @@
   pages={379--388},
   year={1985},
   publisher={OLDENBOURG WISSENSCHAFTSVERLAG},
-  doi={https://doi.org/10.1007/978-1-4612-5769-1}
+  doi={10.1007/978-1-4612-5769-1}
 }
 
 @article{glynn2010introduction,
@@ -159,7 +159,7 @@
   author={Chernozhukov, Victor and Chetverikov, Denis and Demirer, Mert and Duflo, Esther and Hansen, Christian and Newey, Whitney and Robins, James},
   year={2018},
   publisher={Oxford University Press Oxford, UK},
-  doi={https://doi.org/10.1111/ectj.12097}
+  doi={10.1111/ectj.12097}
 }
 
 @software{malenica2022tmle3mopttx,
@@ -179,7 +179,7 @@
[…]
