+++ paper/paper.bib
@@ -151,7 +151,7 @@
   pages={36--56},
   year={2010},
   publisher={Cambridge University Press},
-  doi={doi:10.1093/pan/mpp036}
+  doi={10.1093/pan/mpp036}
 }
 
 @misc{chernozhukov2018double,
