### 2025-08-15 #8446 @jbytecode
@olivierlabayle - The paper and the bibliography look better than ever. Thank you.

Could you please

- Create a tagged release of the software repository. This tagged release will hold the reviewed software.
- Register a zenodo archive using the created tagged release. Obtain a DOI.
- In Zenodo archive the title of the and the list of authors (and their DOIs) should match those in the submission.

After applying those steps, please provide me the tagged release (e.g. v.1.2.3), the archive URL, and the Zenodo DOI here.

Thank you in advance.
