+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v0.20",
   "v0.19",
   "v0.18",
   "v0.17",
@@ -15,5 +16,5 @@ var DOC_VERSIONS = [
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.19.0";
+var DOCUMENTER_NEWEST = "v0.20.0";
 var DOCUMENTER_STABLE = "stable";
