+++ versions.js
@@ -16,5 +16,5 @@ var DOC_VERSIONS = [
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.20.0";
+var DOCUMENTER_NEWEST = "v0.20.1";
 var DOCUMENTER_STABLE = "stable";
