### 2025-08-15 #8446 @olivierlabayle
@jbytecode I've addressed the changes in the title, made a release and a zenodo archive:

- zenodo link: https://zenodo.org/records/16884264
- doi: https://doi.org/10.5281/zenodo.16884264

Please let me know if anything is missing.
