### 2025-08-17 #8446 @jbytecode
@olivierlabayle - Looks good to me. Now I'm recommending an acceptance. Our editor will make the final decision. Thank you!
