### 2025-08-17 #8446 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
